crop-disease-detector/requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
